% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.5ebastian}
\alias{ds.5ebastian}
\title{Fix problem of multiple players named "Sebastian Aho" etc.
imho just including their real-life middle initials is a better solution than "5ebastian"
This function to be updated over time as they move on from their rookie teams}
\usage{
ds.5ebastian(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
the data.frame but with Sebastians Aho etc. disambiguated
}
\description{
Fix problem of multiple players named "Sebastian Aho" etc.
imho just including their real-life middle initials is a better solution than "5ebastian"
This function to be updated over time as they move on from their rookie teams
}
