% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.are_goalie}
\alias{ds.are_goalie}
\title{Are Goalie
\code{are_goalie()} wraps \code{\link{ds.find_goalie}()} to identify goalies}
\usage{
ds.are_goalie(on_df, roster)
}
\arguments{
\item{on_df}{data.frame of who is on ice for the team}

\item{roster}{data.frame recording the team's roster}
}
\value{
vector of goalies
}
\description{
Are Goalie
\code{are_goalie()} wraps \code{\link{ds.find_goalie}()} to identify goalies
}
