% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.are_on}
\alias{ds.are_on}
\title{Are On
\code{are_on()} wraps \code{\link{is_on}()} to identify who all is on ice for each event}
\usage{
ds.are_on(shifts, new_pbp, venue)
}
\arguments{
\item{shifts}{data.frame of shifts}

\item{new_pbp}{data.frame of play-by-play events}

\item{venue}{character. \code{"home"} or \code{"away"}.}
}
\value{
a data.frame of player on-ice statuses by event
}
\description{
Are On
\code{are_on()} wraps \code{\link{is_on}()} to identify who all is on ice for each event
}
