% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\name{ds.clean.nums}
\alias{ds.clean.nums}
\title{Clean Nums
\code{clean_nums()} returns a list of player number identifiers for a given event description}
\usage{
ds.clean.nums(x)
}
\arguments{
\item{x}{event descriptions}
}
\value{
a list of player number IDs
}
\description{
Clean Nums
\code{clean_nums()} returns a list of player number identifiers for a given event description
}
