% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.compile_games}
\alias{ds.compile_games}
\title{Compile Games}
\usage{
ds.compile_games(games = ds.all_games, season = "20182019",
  pause = 1, try_tolerance = 3, agents = ds.user_agents[[1]],
  presave = TRUE, use_presaved = TRUE, data_dir = "nhlgamedata")
}
\arguments{
\item{games}{numeric vector. The game IDs. Defaults to all regular-season and playoff game IDs.}

\item{season}{character. The season. Default \code{"20182019"}.}

\item{pause}{numeric. Seconds to pause between games. Default \code{1}.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character vector. HTTP user agent strings. Defaults to a single sensible string.}

\item{presave}{logical. Save scraped games for later retrieval? Default \code{TRUE}.}

\item{use_presaved}{logical. Reload any presaved games instead of downloading? Default \code{TRUE}.}

\item{data_dir}{character. Path to saved games. Default "nhlgamedata".}
}
\value{
A list object containing c([[1]] = PBP, [[2]] = Roster, [[3]] = Shifts)
}
\description{
\code{compile_games()} collects, parses and compiles all game data corresponding to a given vector of game IDs and season
}
\details{
This is the principal function you should find yourself calling.
}
