% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.compile_nested_games}
\alias{ds.compile_nested_games}
\title{Compile Nested Games
\code{compile_nested_games()} parses and compiles all game data from a list of scraped game results}
\usage{
ds.compile_nested_games(nested_games)
}
\arguments{
\item{nested_games}{list. The output of \code{\link{ds.scrape_nested_games}}.}
}
\value{
A list object containing c([[1]] = PBP, [[2]] = Roster, [[3]] = Shifts)
}
\description{
Compile Nested Games
\code{compile_nested_games()} parses and compiles all game data from a list of scraped game results
}
