% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\name{ds.find_goalie}
\alias{ds.find_goalie}
\title{Find Goalie
\code{find_goalie()} returns a vector containing all goaltenders in a given player vector}
\usage{
ds.find_goalie(players, roster)
}
\arguments{
\item{players}{vector of players.}

\item{roster}{data.frame. The roster}
}
\value{
a vector of goalies.
}
\description{
Find Goalie
\code{find_goalie()} returns a vector containing all goaltenders in a given player vector
}
