% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\name{ds.fix_names}
\alias{ds.fix_names}
\title{Fix Names
\code{fix_names()} returns a vector of player names corrected for multiple spelling variants
Some more of this work is handled in the \code{salo} package. Eventual unification is a desideratum.}
\usage{
ds.fix_names(name_vect)
}
\arguments{
\item{name_vect}{character vector. Player names.}
}
\value{
a vector of player names, standardized.
}
\description{
Fix Names
\code{fix_names()} returns a vector of player names corrected for multiple spelling variants
Some more of this work is handled in the \code{salo} package. Eventual unification is a desideratum.
}
