% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get.R
\name{ds.get_all_players}
\alias{ds.get_all_players}
\title{Get all players}
\usage{
ds.get_all_players(season = "20182019", try_tolerance = 3,
  agents = ds.user_agents[[1]])
}
\arguments{
\item{season}{character. Default \code{20182019}.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a JSON list object
}
\description{
Get all players
}
