% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get.R
\name{ds.get_highlights}
\alias{ds.get_highlights}
\title{Get Highlights}
\usage{
ds.get_highlights(season = "20182019", game_id = "20001",
  try_tolerance = 3, agents = ds.user_agents[[1]])
}
\arguments{
\item{season}{character. The season.  Default \code{"20182019"}.}

\item{game_id}{character. The game id. Default \code{"20001"}.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a JSON list object
}
\description{
\code{get_highlights()} imports the highlights page corresponding to a given year and game ID
}
