% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get.R
\name{ds.get_player_profile}
\alias{ds.get_player_profile}
\title{Get player Profile
\code{get_team_profile()} imports the team profile page corresponding to a given team ID}
\usage{
ds.get_player_profile(player_id = 8477474, try_tolerance = 3,
  agents = ds.user_agents[[1]])
}
\arguments{
\item{player_id}{numeric. Default \code{8477474} (Madison Bowey).}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a JSON list object
}
\description{
Get player Profile
\code{get_team_profile()} imports the team profile page corresponding to a given team ID
}
