% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get.R
\name{ds.get_team_profile}
\alias{ds.get_team_profile}
\title{Get Team Profile
\code{get_team_profile()} imports the team profile page corresponding to a given team ID}
\usage{
ds.get_team_profile(team_id = 1, try_tolerance = 3,
  agents = ds.user_agents[[1]])
}
\arguments{
\item{team_id}{numeric. Default \code{1} (NJD).}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a JSON list object
}
\description{
Get Team Profile
\code{get_team_profile()} imports the team profile page corresponding to a given team ID
}
