% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_mung.R
\name{ds.mung_names}
\alias{ds.mung_names}
\title{pick one version of certain guys' names and stick with it}
\usage{
ds.mung_names(name_vector)
}
\arguments{
\item{name_vector}{a vector of names}
}
\value{
a vector of the same length but with each individual allowed only one version of their name
}
\description{
pick one version of certain guys' names and stick with it
}
