% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.name_player}
\alias{ds.name_player}
\title{Name Player
\code{name_player()} looks up names for coded players in a play-by-play object under construction}
\usage{
ds.name_player(player, full_pbp, roster)
}
\arguments{
\item{player}{character. which player.}

\item{full_pbp}{data.frame. The play-by-play.}

\item{roster}{data.frame. The team's roster.}
}
\value{
a character vector.
}
\description{
Name Player
\code{name_player()} looks up names for coded players in a play-by-play object under construction
}
