% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_date}
\alias{ds.parse_date}
\title{Parse Date}
\usage{
ds.parse_date(x)
}
\arguments{
\item{x}{a single entry.}
}
\value{
a data.frame
}
\description{
\code{parse_date()} parses a single date from the Schedule JSON object
}
\details{
\code{parse_date()} uses an inner function \code{parse_game()}
}
