% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_event}
\alias{ds.parse_event}
\title{Parse PBP Event}
\usage{
ds.parse_event(x)
}
\arguments{
\item{x}{a single entry.}
}
\value{
a data.frame
}
\description{
\code{parse_event()} parses a single event from the PBP JSON object
}
