% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_game}
\alias{ds.parse_game}
\title{Parse Game}
\usage{
ds.parse_game(x)
}
\arguments{
\item{x}{a single entry.}
}
\value{
a data.frame
}
\description{
\code{parse_game()} parses a single game from the Schedule >> Date JSON object
}
\details{
\code{parse_game()} is an inner function for \code{parse_date()}
}
