% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_highlight}
\alias{ds.parse_highlight}
\title{Parse Highlight
\code{parse_highlight()} parses a single highlight from the Highlights JSON object}
\usage{
ds.parse_highlight(x)
}
\arguments{
\item{x}{a single entry.}
}
\value{
a data.frame
}
\description{
Parse Highlight
\code{parse_highlight()} parses a single highlight from the Highlights JSON object
}
