% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_player}
\alias{ds.parse_player}
\title{Parse Player
\code{parse_player()} parses a single player from the PBP JSON object and returns a data frame}
\usage{
ds.parse_player(x)
}
\arguments{
\item{x}{a single entry.}
}
\value{
a data.frame
}
\description{
Parse Player
\code{parse_player()} parses a single player from the PBP JSON object and returns a data frame
}
