% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_shift}
\alias{ds.parse_shift}
\title{Parse Shift
\code{parse_shift()} parses a single shift from the Shifts JSON object}
\usage{
ds.parse_shift(x)
}
\arguments{
\item{x}{A single entry}
}
\value{
a data.frame
}
\description{
Parse Shift
\code{parse_shift()} parses a single shift from the Shifts JSON object
}
