% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_parse.R
\name{ds.parse_shifts}
\alias{ds.parse_shifts}
\title{Parse Shifts
\code{parse_shifts()} returns a matrix containing shift information for a single player}
\usage{
ds.parse_shifts(player, venue, inner, outer)
}
\arguments{
\item{player}{the player}

\item{venue}{character. \code{"Home"} or \code{"Away"}.}

\item{inner}{inner (second) element of the output of \code{ds.get_shifts}.}

\item{outer}{outer  (first) element of the output of \code{ds.get_shifts}.}
}
\value{
a matrix containing shift info
}
\description{
Parse Shifts
\code{parse_shifts()} returns a matrix containing shift information for a single player
}
