% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\docType{data}
\name{ds.pbp_colnames}
\alias{ds.pbp_colnames}
\title{Play-by-play column names}
\format{character vector}
\usage{
ds.pbp_colnames
}
\description{
Play-by-play column names
}
\keyword{datasets}
