% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_pull.R
\name{ds.pull_JSON}
\alias{ds.pull_JSON}
\title{Don't Repeat Yourself when pulling JSON}
\usage{
ds.pull_JSON(url, agents, try_tolerance = 3, ref = "nhl.com",
  clean = FALSE)
}
\arguments{
\item{url}{character.        A URL.}

\item{agents}{character vector. One or more HTTP user agent strings.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{ref}{character.        A referer. Default \code{"nhl.com"}.}

\item{clean}{logical. Clean certain special characters from the JSON? Default \code{FALSE}.}
}
\value{
a JSON object or \code{NULL}
}
\description{
Don't Repeat Yourself when pulling JSON
}
