% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_pull.R
\name{ds.pull_text}
\alias{ds.pull_text}
\title{Don't Repeat Yourself when pulling text}
\usage{
ds.pull_text(url, agents, try_tolerance = 3, ref = "nhl.com")
}
\arguments{
\item{url}{character.        A URL.}

\item{agents}{character vector. One or more HTTP user agent strings.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{ref}{character.        A referer. Default \code{"nhl.com"}.}
}
\value{
a nice HTML thingy or \code{NULL}
}
\description{
Don't Repeat Yourself when pulling text
}
