% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_pull.R
\name{ds.pull_URL}
\alias{ds.pull_URL}
\title{Don't Repeat Yourself}
\usage{
ds.pull_URL(url, agents, ref = "nhl.com")
}
\arguments{
\item{url}{character.        A URL.}

\item{agents}{character vector. One or more HTTP user agent strings.}

\item{ref}{character.        A referer. Default \code{"nhl.com"}.}
}
\value{
output of \code{RCurl::\link[RCurl]{getURL}}.
}
\description{
\code{pull_URL()} is a one-line wrapper around \code{RCurl::getURL()} with repeatedly used options
}
