% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_scrape.R
\name{ds.scrape_game}
\alias{ds.scrape_game}
\title{Scrape Game}
\usage{
ds.scrape_game(season = "20182019", game_id = "20001",
  try_tolerance = 3, agents = ds.user_agents[[1]], presave = TRUE,
  use_presaved = TRUE, data_dir = "nhlgamedata")
}
\arguments{
\item{season}{character. The season. Default \code{"20182019"}.}

\item{game_id}{character. Game ID. Default \code{20001}.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character vector. HTTP user agent strings. Defaults to a single sensible string.}

\item{presave}{logical. Save scraped games for later retrieval? Default \code{TRUE}.}

\item{use_presaved}{logical. Reload any presaved games instead of downloading? Default \code{TRUE}.}

\item{data_dir}{character. Path to saved games. Default "nhlgamedata".}
}
\value{
A list object containing c([[1]] = PBP, [[2]] = Shifts, [[3]] = Highlights)
}
\description{
\code{scrape_game()} collects the data for a game corresponding to a given season and game ID
}
