% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_scrape.R
\name{ds.scrape_nested_games}
\alias{ds.scrape_nested_games}
\title{Scrape Multiple Games
\code{scrape_nested_games()} collects all game data corresponding to a given vector of game IDs and season}
\usage{
ds.scrape_nested_games(games = ds.all_games, season = "20182019",
  pause = 1, try_tolerance = 3, agents = ds.user_agents[[1]],
  presave = TRUE, use_presaved = TRUE, data_dir = "nhlgamedata")
}
\arguments{
\item{games}{numeric vector. The game IDs. Defaults to all regular-season and playoff game IDs.}

\item{season}{character. The season. Default \code{"20182019"}.}

\item{pause}{numeric. Seconds to pause between games. Default \code{1}.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character vector. HTTP user agent strings. Defaults to a single sensible string.}

\item{presave}{logical. Save scraped games for later retrieval? Default \code{TRUE}.}

\item{use_presaved}{logical. Reload any presaved games instead of downloading? Default \code{TRUE}.}

\item{data_dir}{character. Path to saved games. Default "nhlgamedata".}
}
\value{
A list of several five-element lists, one for each game
}
\description{
Scrape Multiple Games
\code{scrape_nested_games()} collects all game data corresponding to a given vector of game IDs and season
}
