% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_scrape.R
\name{ds.scrape_player_profile}
\alias{ds.scrape_player_profile}
\title{Scrape Player Profile
\code{scrape_player_profile()} collects and parses the data for a player corresponsing to a given player ID}
\usage{
ds.scrape_player_profile(player_id = 8477474, try_tolerance = 3,
  agents = ds.user_agents[[1]])
}
\arguments{
\item{player_id}{numeric. Default \code{8477474} (Madison Bowey).}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a data.frame
}
\description{
Scrape Player Profile
\code{scrape_player_profile()} collects and parses the data for a player corresponsing to a given player ID
}
