% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_scrape.R
\name{ds.scrape_schedule}
\alias{ds.scrape_schedule}
\title{Scrape Schedule
\code{scrape_schedule()} collects and parses the schedule data for a range corresponding to a given start and end date}
\usage{
ds.scrape_schedule(start = "2018-10-01", end = "2019-06-30",
  try_tolerance = 3, agents = ds.user_agents[[1]])
}
\arguments{
\item{start}{character or date object. Default \code{"2018-10-01"}.}

\item{end}{character or date object. Default \code{"2019-06-30"}.}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a data.frame
}
\description{
Scrape Schedule
\code{scrape_schedule()} collects and parses the schedule data for a range corresponding to a given start and end date
}
