% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_scrape.R
\name{ds.scrape_team_profile}
\alias{ds.scrape_team_profile}
\title{Scrape Team Profile
\code{scrape_team_profile()} collects and parses the data for a team corresponsing to a given team ID}
\usage{
ds.scrape_team_profile(team_id, try_tolerance = 3,
  agents = ds.user_agents[[1]])
}
\arguments{
\item{team_id}{numeric. Default \code{1} (NJD).}

\item{try_tolerance}{numeric. Number of times to bang head against wall. Default \code{3}.}

\item{agents}{character. Agent string. Default should work fine.}
}
\value{
a data.frame
}
\description{
Scrape Team Profile
\code{scrape_team_profile()} collects and parses the data for a team corresponsing to a given team ID
}
