% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\name{ds.seconds_from_ms}
\alias{ds.seconds_from_ms}
\title{Seconds from MS
\code{seconds_from_ms()} returns a numeric vector of representation in seconds of a given vector in M:S format}
\usage{
ds.seconds_from_ms(ms)
}
\arguments{
\item{ms}{string. Minutes and seconds.}
}
\value{
a number of seconds.
}
\description{
Seconds from MS
\code{seconds_from_ms()} returns a numeric vector of representation in seconds of a given vector in M:S format
}
