% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.shift_sum}
\alias{ds.shift_sum}
\title{summary of shifts
This is just a helper function used to reduce repetition within a function above.}
\usage{
ds.shift_sum(shifts, type = "ON")
}
\arguments{
\item{shifts}{data.frame. Shift endpoint info.}

\item{type}{character. \code{"ON"} or \code{"OFF"}. Default \code{"ON"}.}
}
\description{
summary of shifts
This is just a helper function used to reduce repetition within a function above.
}
