% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\docType{data}
\name{ds.user_agents}
\alias{ds.user_agents}
\title{A set of sensible HTTP user agent strings}
\format{character vector}
\usage{
ds.user_agents
}
\description{
A set of sensible HTTP user agent strings
}
\keyword{datasets}
