% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\name{ds.who}
\alias{ds.who}
\title{Who
\code{who()} searches a given player ID and returns the player's full name}
\usage{
ds.who(player_id)
}
\arguments{
\item{player_id}{numeric. Player ID.}
}
\value{
full name.
}
\description{
Who
\code{who()} searches a given player ID and returns the player's full name
}
