% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_compile.R
\name{ds.who_is}
\alias{ds.who_is}
\title{Who Is
\code{who_is()} is a helper function used by \code{\link{ds.name_player}()}.}
\usage{
ds.who_is(x, i)
}
\arguments{
\item{x}{data.frame.}

\item{i}{column name.}
}
\value{
\code{x[[i]]} mangled with \code{x$game_id}.
}
\description{
Who Is
\code{who_is()} is a helper function used by \code{\link{ds.name_player}()}.
}
