% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_dryscrape.R
\name{is_on}
\alias{is_on}
\title{Is On
\code{is_on()} returns a numeric vector indicating 1 if a given player is on ice during the event corresponding to the row index in the given PBP pbject}
\usage{
is_on(player, pbp, venue)
}
\arguments{
\item{player}{the player.}

\item{pbp}{data.frame. The play-by-play.}

\item{venue}{character. \code{"home"} oe \code{"away"}.}
}
\value{
a numeric vector.
}
\description{
Is On
\code{is_on()} returns a numeric vector indicating 1 if a given player is on ice during the event corresponding to the row index in the given PBP pbject
}
